Checkout page and it's components rewritten to semantic

+++ src/pages/cart/PurchaseOrder/PurchaseOrder.js
@@ -1,18 +1,30 @@
 import React, { Component } from 'react'
 import PropTypes from "prop-types"
 import { actions } from 'react-redux-form'
-import SummaryTable from "../components/SummaryTable/SummaryTable"
+
 import Shipping from "./components/Shipping"
 import ShippingEdit from "./components/ShippingEdit"
 import ShippingQuote from "./components/ShippingQuote"
 import Payment from "./components/Payment"
-import CartItemSummary from './components/CartItemSummary'
-import { Container, Menu, Header, Button, Icon } from "semantic-ui-react"
+import { Container, Menu, Header, Button, Icon, Grid, GridColumn, GridRow, Segment, Divider } from "semantic-ui-react"
+import styled from 'styled-components'
 import Spinner from '../../../components/Spinner/Spinner'
 import "./PurchaseOrder.scss"
-import {FormattedMessage} from 'react-intl'
-import {checkToken} from "../../../utils/auth"
+import { FormattedMessage, injectIntl } from 'react-intl'
+import { checkToken } from "../../../utils/auth"
 import Router from 'next/router'
+
+import CartItemSummary from '~/components/summary/CartItemSummary'
+import Summary from '~/components/summary/Summary'
+
+import confirm from '../../../components/Confirmable/confirm';
+
+
+const RelaxedGrid = styled(Grid)`
+  margin-top 1.5rem !important;
+  padding-bottom: 50px !important;
+`
+
 
 class PurchaseOrder extends Component {
   //TODO: maybe move internal state to redux? decide it later
@@ -20,8 +32,10 @@ class PurchaseOrder extends Component {
     selectedAddress: {},
     selectedPayment: {},
     isShippingEdit: false,
-    isNewAddress: "isNew",
-    shippingQuotes: []
+    isNewAddress: true,
+    savedShippingPreferences: true,
+    shippingQuotes: [],
+    selectedShippingQuote: null
   }
 
   constructor(props) {
@@ -29,43 +43,48 @@ class PurchaseOrder extends Component {
     this.deleteCart = this.deleteCart.bind(this);
   }
 
-  componentDidMount(){
+  componentDidMount() {
     this.props.getCart()
     this.props.getDeliveryAddresses()
     this.props.getPayments()
   }
 
   handleIsEdit = (value) => {
-    const {selectedAddress} = this.state;
-    this.setState({isNewAddress: value});
+    this.setState({ isShippingEdit: value });
 
-    value === "isNew"
-    ? this.props.dispatch(actions.reset('forms.shippingEdit'))
-    : this.props.dispatch(actions.merge('forms.shippingEdit', {
-      firstName: selectedAddress["first name"],
-      lastName: selectedAddress["last name"],
-      address: {
-        streetAddress: selectedAddress.address.streetAddress,
-        city: selectedAddress.address.city,
-        province: selectedAddress.address.province.name
-      },
-      zipCode: selectedAddress.address.zip.zip,
-      email: selectedAddress.email,
-      phoneNumber: selectedAddress["phone number"]
-  }));
+    // value === "isNew"
+    //   ? this.props.dispatch(actions.reset('forms.shippingEdit'))
+    //   : this.props.dispatch(actions.merge('forms.shippingEdit', {
+    //     firstName: selectedAddress["first name"],
+    //     lastName: selectedAddress["last name"],
+    //     address: {
+    //       streetAddress: selectedAddress.address.streetAddress,
+    //       city: selectedAddress.address.city,
+    //       province: selectedAddress.address.province.name
+    //     },
+    //     zipCode: selectedAddress.address.zip.zip,
+    //     email: selectedAddress.email,
+    //     phoneNumber: selectedAddress["phone number"]
+    //   }));
+  }
+
+  handleQuoteSelect = (index) => {
+    // Redux, cause it should change prize 
+    let { shippingQuoteSelected, shippingQuotes } = this.props
+    shippingQuoteSelected({ index, quote: shippingQuotes[index] })
   }
 
   getAddress = (selectedAddressId) => {
-    const {deliveryAddresses} = this.props;
+    const { deliveryAddresses } = this.props;
     const selectedAddress = deliveryAddresses.find(i => i.id === selectedAddressId);
-    this.setState({selectedAddress});
+    this.setState({ selectedAddress });
     this.getShippingQuotes(selectedAddress);
   }
 
   getPayment = (selectedPaymentId) => {
-    const {payments} = this.props;
+    const { payments } = this.props;
     const selectedPayment = payments.find(i => i.id === selectedPaymentId);
-    this.setState({selectedPayment});
+    this.setState({ selectedPayment });
   }
 
   getShippingQuotes = (selectedAddress) => {
@@ -73,67 +92,36 @@ class PurchaseOrder extends Component {
     this.props.getShippingQuotes(1, selectedAddress.address.zip.zip);
   }
 
-  toggleShippingEdit = () => {
+  toggleRadio = (name = 'isNewAddress') => {
     this.setState(prevState => ({
-      isShippingEdit: !prevState.isShippingEdit
-    }));
+      [name]: !prevState[name]
+    }))
   }
 
-  //TODO:: same function in Shopping cart, define it just at one place
-  renderSummary() {
-    const {cartItems, totalPrice} = this.props.cart;
-    let subtotal = 0;
-    for (let i = 0; i < cartItems.length; i++) {
-      subtotal+= (cartItems[i].quantity * cartItems[i].productOffer.product.packagingSize * cartItems[i].productOffer.pricing.price.amount)
+  deleteCart(id) {
+    if (checkToken(this.props)) return
+    let { cart } = this.props
+
+
+    let { formatMessage } = this.props.intl
+
+    if (cart.cartItems.length === 1) {
+      return confirm(
+        formatMessage(({
+          id: 'order.deleteHeader',
+          defaultMessage: 'Delete Order'
+        })),
+        formatMessage(({
+          id: 'order.deleteBody',
+          defaultMessage: 'You are about to delete last item of order. Doing so will redirect you to Shopping cart. Do you wish to continue?'
+        }))
+      ).then(() => {
+        this.props.deleteCart()
+        Router.push('/cart')
+      })
+    } else {
+      this.props.deleteCartItem(id)
     }
-    let currencySymbol = cartItems.length ? cartItems[0].productOffer.pricing.price.currency.symbol : '$'
-    return (
-      <table>
-        <tbody>
-          <tr>
-              <td>
-                  <FormattedMessage
-                      id='cart.subtotal'
-                      defaultMessage='Subtotal'
-                  />
-              </td>
-              <td>{currencySymbol}{subtotal.formatMoney(0)}</td>
-          </tr>
-          <tr>
-              <td>
-                  <FormattedMessage
-                    id='cart.estimatedShipping'
-                    defaultMessage='Estimated Shipping'
-                  />
-              </td>
-              <td></td>
-          </tr>
-          <tr>
-              <td>
-                  <FormattedMessage
-                    id='cart.estimatedTax'
-                    defaultMessage='Estimated Tax'
-                  />
-              </td>
-              <td></td>
-          </tr>
-          <tr>
-              <td>
-                  <FormattedMessage
-                    id='cart.total'
-                    defaultMessage='Total'
-                  />
-              </td>
-              <td>{currencySymbol}{totalPrice.formatMoney(0)}</td>
-          </tr>
-        </tbody>
-      </table>
-    )
-  }
-
-  deleteCart() {
-      if (checkToken(this.props)) return;
-      this.props.deleteCart();
   }
 
   handlePurchase() {
@@ -141,18 +129,10 @@ class PurchaseOrder extends Component {
   }
 
   render() {
-    const {cart, deliveryAddresses, payments, dispatch, deleteCart, cartIsFetching, postNewDeliveryAddress, putDeliveryAddressEdit, shippingQuotes} = this.props;
+    const { cart, deliveryAddresses, payments, dispatch, deleteCart, cartIsFetching, postNewDeliveryAddress, putDeliveryAddressEdit, shippingQuotes } = this.props;
     if (cartIsFetching) return <Spinner />
-    let index = 0;
-    const itemContent = cart.cartItems.map(cartItem => {
-      return (
-      <CartItemSummary
-        deleteCart={deleteCart}
-        cartItem={cartItem}
-        key={cartItem.productOffer.id}
-        itemIndex={++index}
-      />)
-    });
+
+
     return (
       <div className="app-inner-main">
         <div className="header-top">
@@ -161,7 +141,7 @@ class PurchaseOrder extends Component {
               <Menu.Item header>
                 <Header as='h1' size='medium'>
                   <FormattedMessage id='cart.purchaseOrder'
-                                    defaultMessage='PURCHASE ORDER' />
+                    defaultMessage='PURCHASE ORDER' />
                 </Header>
               </Menu.Item>
 
@@ -170,62 +150,133 @@ class PurchaseOrder extends Component {
                   <Button icon basic labelPosition='left' onClick={() => { Router.push('/marketplace/all') }}>
                     <Icon name='chevron left' />
                     <FormattedMessage id='cart.backToProductOfferings'
-                                      defaultMessage='Back to Product Offerings' />
+                      defaultMessage='Back to Product Offerings' />
                   </Button>
                 </Menu.Item>
               </Menu.Menu>
             </Menu>
           </Container>
         </div>
-        <div className="shopping-cart checkout">
-          <div className="shopping-cart-body">
 
-          <div>
-            {this.state.isShippingEdit ? <ShippingEdit
-                toggleShippingEdit={this.toggleShippingEdit}
+        <RelaxedGrid centered className='purchase-order'>
+
+          <GridColumn computer={9}>
+
+            {this.state.isShippingEdit ?
+              <ShippingEdit
+                savedShippingPreferences={this.state.savedShippingPreferences}
+                toggleRadio={(name) => this.toggleRadio(name)}
                 selectedAddress={this.state.selectedAddress}
                 isNewAddress={this.state.isNewAddress}
                 handleIsEdit={this.handleIsEdit}
                 postNewDeliveryAddress={postNewDeliveryAddress}
                 putDeliveryAddressEdit={putDeliveryAddressEdit}
               />
-              : <Shipping
-              deliveryAddresses={deliveryAddresses}
-              dispatch={dispatch}
-              toggleShippingEdit={this.toggleShippingEdit}
-              getAddress={this.getAddress}
-              selectedAddress={this.state.selectedAddress}
-              />}
-            <ShippingQuote
-              selectedAddress={this.state.selectedAddress}
-              shippingQuotes={shippingQuotes}
-              shippingQuotesAreFetching={this.props.shippingQuotesAreFetching}
-              />
-            <Payment
-              dispatch={dispatch}
-              selectedAddress={this.state.selectedAddress}
-              selectedPayment={this.state.selectedPayment}
-              payments={payments}
-              getPayment={this.getPayment}
-              />
+              :
+              <>
+                <Shipping
+                  deliveryAddresses={deliveryAddresses}
+                  dispatch={dispatch}
+                  handleIsEdit={this.handleIsEdit}
+                  toggleRadio={(name) => this.toggleRadio(name)}
+                  getAddress={this.getAddress}
+                  selectedAddress={this.state.selectedAddress}
+                />
+              </>
+            }
 
-          </div>
-          <div className="summary-tables">
-            <SummaryTable title="Your Order">
-              {itemContent}
-            </SummaryTable>
-            <SummaryTable title="Summary" hasButton={<FormattedMessage id='cart.placeOrder' defaultMessage='Place Order' />} handleContinue={this.handlePurchase}>
-              {this.renderSummary()}
-            </SummaryTable>
-          </div>
-          </div>
-        </div>
+
+
+
+            <Segment>
+              <Grid className='bottom-padded'>
+                <GridRow className='header'>
+                  <GridColumn>
+                    <Header as='h2'>
+                      <FormattedMessage
+                        id='cart.2freightSelection'
+                        defaultMessage='2. Freight Selection'
+                      />
+                    </Header>
+                  </GridColumn>
+                </GridRow>
+
+                <ShippingQuote
+                  //TODO, change when backend provides this info
+                  currency={{ code: 'USD' }}
+                  selectedShippingQuote={this.props.cart.selectedShipping}
+                  handleQuoteSelect={this.handleQuoteSelect}
+                  selectedAddress={this.state.selectedAddress}
+                  shippingQuotes={shippingQuotes}
+                  shippingQuotesAreFetching={this.props.shippingQuotesAreFetching}
+                />
+              </Grid>
+            </Segment>
+
+
+
+            <Segment>
+              <Grid className='bottom-bottom-padded'>
+                <GridRow className='header'>
+                  <GridColumn>
+                    <Header as='h2'>
+                      <FormattedMessage
+                        id='cart.3payment'
+                        defaultMessage='3. Payment'
+                      />
+                    </Header>
+                  </GridColumn>
+                </GridRow>
+
+                <Payment
+                  dispatch={dispatch}
+                  selectedAddress={this.state.selectedAddress}
+                  selectedPayment={this.state.selectedPayment}
+                  payments={payments}
+                  getPayment={this.getPayment}
+                />
+              </Grid>
+            </Segment>
+
+
+          </GridColumn>
+
+          <GridColumn computer={5}>
+
+
+            <CartItemSummary
+              cartItems={cart.cartItems}
+              deleteCart={this.deleteCart}
+            />
+
+            <Summary
+              additionalContent={
+                <GridRow centered>
+                  <GridColumn>
+                    <Button fluid primary onClick={this.handleContinue}>
+                      <FormattedMessage
+                        id='cart.placeOrder'
+                        defaultMessage='Place Order1' />
+                    </Button>
+                  </GridColumn>
+                </GridRow>
+              }
+              handleContinue={this.handlePurchase}
+              cart={cart}
+              totalPrice={this.props.cart.totalPrice}
+            />
+
+          </GridColumn>
+
+
+        </RelaxedGrid>
+
       </div>
     )
   }
 }
 
-export default PurchaseOrder;
+export default injectIntl(PurchaseOrder)
 
 PurchaseOrder.propTypes = {
   cartItem: PropTypes.object,
